chore: reorganize state specs
